# golang/docs/swagger.yaml
basePath: /api
definitions:
...
  api.Attr:
    properties:
      context:
        description: 内容
        example: '{"content":"请问当前时间是几点？","role":"user"}'
        type: string
      id:
        description: Example ID
        example: 0
        type: integer
      name:
        description: Tab
        example: 商城商品
        type: string
      type:
        description: 1 chat completion 2 img
        enum:
        - 1
        - 2
        example: 1
        type: integer
    type: object
  api.ChatAttr:
    properties:
      context:
        items:
          $ref: '#/definitions/api.ChatCompletionMessage'
        type: array
      id:
        type: integer
      name:
        description: Tab
        example: 商城商品
        type: string
      type:
        description: 1 chat completion 2 img
        example: 1
        type: integer
    type: object
  api.ChatCompletionMessage:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
  api.ChatCompletionResponse:
    properties:
      context:
        type: string
    type: object
...
  api.UserMessage:
    properties:
      content:
        type: string
    type: object
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  common.SuccResponse:
    properties:
      succ:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a API documentation for OPENAI.
  title: OPENAI
  version: "1.0"
paths:
...
  /v1/apps/{app_id}/attrs:
    get:
      description: Get all attrs
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Attr'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get all Attrs
      tags:
      - Attr
    post:
      consumes:
      - application/json
      description: 用户通用属性创建，如果是文字和图片场景，请参考 对话Attr创建 和 图片Attr创建
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      - description: Attr data
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/api.Attr'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Attr'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 通用Attr创建
      tags:
      - Attr
  /v1/apps/{app_id}/attrs/{attr_id}:
    get:
      description: Get Attr by ID
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      - description: Attr ID
        in: path
        name: attr_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Attr'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Attr by ID
      tags:
      - Attr
  /v1/apps/{app_id}/chat_attrs:
    post:
      consumes:
      - application/json
      description: 使用提供的数据创建新的 对话Attr
      parameters:
      - description: 应用ID
        in: path
        name: app_id
        required: true
        type: integer
      - description: 更新的应用数据
        in: body
        name: appData
        required: true
        schema:
          $ref: '#/definitions/api.ChatAttr'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Attr'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 对话Attr创建
      tags:
      - Attr
  /v1/apps/{app_id}/chat_attrs/{attr_id}:
    put:
      consumes:
      - application/json
      description: 使用提供的数据创建新的 对话Attr
      parameters:
      - description: 应用ID
        in: path
        name: app_id
        required: true
        type: integer
      - description: AttrID
        in: path
        name: attr_id
        required: true
        type: integer
      - description: 更新的应用数据
        in: body
        name: appData
        required: true
        schema:
          $ref: '#/definitions/api.ChatAttr'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Attr'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 对话Attr创建
      tags:
      - Attr
  /v1/apps/{appID}/attrs/{attrID}:
    delete:
      consumes:
      - application/json
      description: Delete an attribute by providing the app ID, attribute ID, and
        updated attribute information
      parameters:
      - description: App ID
        format: int32
        in: path
        name: appID
        required: true
        type: integer
      - description: Attribute ID
        format: int32
        in: path
        name: attrID
        required: true
        type: integer
      - description: Updated attribute information
        in: body
        name: updatedAttr
        required: true
        schema:
          $ref: '#/definitions/api.Attr'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/common.SuccResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete an attribute
      tags:
      - Attr
    put:
      consumes:
      - application/json
      description: Update an attribute by providing the app ID, attribute ID, and
        updated attribute information
      parameters:
      - description: App ID
        format: int32
        in: path
        name: appID
        required: true
        type: integer
      - description: Attribute ID
        format: int32
        in: path
        name: attrID
        required: true
        type: integer
      - description: Updated attribute information
        in: body
        name: updatedAttr
        required: true
        schema:
          $ref: '#/definitions/api.Attr'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/common.SuccResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update an attribute
      tags:
      - Attr
...
  /v1/gpt/{attrID}/chat-completion:
    post:
      consumes:
      - application/json
      description: Generate chat completion text based on input messages.
      parameters:
      - description: Attr ID
        in: path
        name: attrID
        required: true
        type: integer
      - description: User Messages
        in: body
        name: userMessage
        required: true
        schema:
          $ref: '#/definitions/api.UserMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChatCompletionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Generate Chat Completion
      tags:
      - ChatGpt
...
swagger: "2.0"